Created a new file for menu options
We are integration 3 types of searches ( by event name, by venue and by date ). Likewise 3 types of Sorting. Hence created a new header file to store the date. Additionally to prevent code repetition for lower case functions ( to check equality of search and conditions in sort ), we added a new function toLowerCase() to deal with this.